Cleanup movement

+++ robot/src/main/java/group14/robot/implementation/Movement.java
@@ -16,7 +16,7 @@ public class Movement extends UnicastRemoteObject implements IMovement {
     // All values below are in mm
     private static final double MARGIN_OF_ERROR = -0.05;
     private static final double MARGIN_OF_ERROR_TURN = 2;
-    private static final double ROBOT_DIAMETER = 114.3+22;
+    private static final double ROBOT_DIAMETER = 114.3 + 22;
     private static final double WHEEL_DIAMETER = 39;
 
 
@@ -63,7 +63,7 @@ public class Movement extends UnicastRemoteObject implements IMovement {
                 break;
 
             case TURN:
-                this.turn((int) instruction.getAmount());
+                this.turn(instruction.getAmount());
                 break;
 
             case DEPOSIT:
@@ -83,7 +83,7 @@ public class Movement extends UnicastRemoteObject implements IMovement {
     @Override
     public void runInstruction(InstructionOld instruction) {
         if (instruction.getAngle() != 0) {
-            this.turn((int) instruction.getAngle());
+            this.turn(instruction.getAngle());
         }
 
         this.forward(instruction.getDistance() * 10);
@@ -99,23 +99,24 @@ public class Movement extends UnicastRemoteObject implements IMovement {
 
     /**
      * calculates required degrees to rotate to move the required distance
+     *
      * @param distance as mm
      */
     @Override
-    public void forward(double distance){
+    public void forward(double distance) {
         distance = distance + (MARGIN_OF_ERROR * distance);
-        double deg = -(360*distance/(WHEEL_DIAMETER * Math.PI));
+        double deg = -(360 * distance / (WHEEL_DIAMETER * Math.PI));
 //        if (distance < 0){
 //            deg -= MARGIN_OF_ERROR;
 //        } else {
 //            deg += MARGIN_OF_ERROR;
 //        }
         this.motorLeft.endSynchronization();
-        this.motorLeft.rotate((int)deg, true);
-        this.motorRight.rotate((int)deg, true);
+        this.motorLeft.rotate((int) deg, true);
+        this.motorRight.rotate((int) deg, true);
         this.motorLeft.endSynchronization();
 
-        while (this.motorLeft.isMoving() || this.motorRight.isMoving()) {}
+        while (this.motorLeft.isMoving() || this.motorRight.isMoving()) { }
     }
 
     @Override
@@ -127,7 +128,6 @@ public class Movement extends UnicastRemoteObject implements IMovement {
     }
 
     /**
-     *
      * @param distance as mm
      */
     @Override
@@ -144,16 +144,17 @@ public class Movement extends UnicastRemoteObject implements IMovement {
     }
 
     @Override
-    public void turn(int degree) {
+    public void turn(double degree) {
         double deg = ((((ROBOT_DIAMETER * Math.PI)) / (WHEEL_DIAMETER * Math.PI)) * 360) * ((double) degree / 360);
-        deg += deg/360*MARGIN_OF_ERROR_TURN;
+        deg += deg / 360 * MARGIN_OF_ERROR_TURN;
+
         int tempSpeed = this.motorLeft.getSpeed();
         this.setSpeedPercentage(25);
 
         this.motorLeft.rotate((int) deg, true);
         this.motorRight.rotate((int) -deg, true);
 
-        while (this.motorLeft.isMoving() || this.motorRight.isMoving()) {}
+        while (this.motorLeft.isMoving() || this.motorRight.isMoving()) { }
 
         this.motorLeft.setSpeed(tempSpeed);
         this.motorRight.setSpeed(tempSpeed);
