+++ robot/src/main/java/group14/robot/interfaces/IMovement.java
@@ -15,6 +15,6 @@ public interface IMovement extends Remote {
     void backward() throws RemoteException;
     void backward(double distance) throws RemoteException;
     void stop() throws RemoteException;
-    void turn(int degree) throws RemoteException;
+    void turn(double degree) throws RemoteException;
     void setSpeedPercentage(double percent) throws RemoteException;
 }
